created a new page to fetch the SSR page
